+++ lib/convenient_service/rspec/helpers/custom/wrap_method/entities/wrapped_method.rb
@@ -35,9 +35,25 @@ module ConvenientService
                   #
                   stack.use(
                     proc do |env|
-                      entity.__send__(method, *env[:args], **env[:kwargs], &env[:block])
-                        .tap { |value| @chain_value = value }
-                        .tap { @chain_arguments = {args: env[:args], kwargs: env[:kwargs], block: env[:block]} }
+                      @chain_arguments = {args: env[:args], kwargs: env[:kwargs], block: env[:block]}
+
+                      ##
+                      # IMPORTANT: Forces Ruby to define `@chain_value` instance variable.
+                      #
+                      # NOTE: If `@chain_value` is still set to `Support::UNDEFINED` after running the begin block - an exception was raised by `chain.next`. See `chain_called?` for more info.
+                      #
+                      @chain_value = Support::UNDEFINED
+
+                      begin
+                        @chain_value = entity.__send__(method, *env[:args], **env[:kwargs], &env[:block])
+                      rescue => exception
+                        @chain_exception = exception
+
+                        ##
+                        # NOTE: `raise` with no args inside `rescue` reraises rescued exception.
+                        #
+                        raise
+                      end
                     end
                   )
                 end
@@ -106,6 +122,16 @@ module ConvenientService
                 raise Errors::ChainAttributePreliminaryAccess.new(attribute: :block) unless chain_called?
 
                 @chain_arguments[:block]
+              end
+
+              ##
+              # @return [StandardError, nil]
+              # @raise [ConvenientService::RSpec::Helpers::Custom::WrapMethod::Errors::ChainAttributePreliminaryAccess]
+              #
+              def chain_exception
+                raise Errors::ChainAttributePreliminaryAccess.new(attribute: :exception) unless chain_called?
+
+                @chain_exception
               end
             end
           end