+++ lib/convenient_service/service/plugins/rescues_result_unhandled_exceptions/commands.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+require_relative "commands/format_backtrace"
+require_relative "commands/format_cause"
+require_relative "commands/format_line"
+
+require_relative "commands/format_exception"