+++ lib/convenient_service/support/dependency_container/errors.rb
@@ -68,6 +68,20 @@ module ConvenientService
             super(message)
           end
         end
+
+        class NotModule < ConvenientService::Error
+          ##
+          # @param klass [Class]
+          # @return [void]
+          #
+          def initialize(klass:)
+            message = <<~TEXT
+              `#{klass.inspect}` is NOT a Module.
+            TEXT
+
+            super(message)
+          end
+        end
       end
     end
   end